requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-config-url==0.1.1
Django==5.1.1
django-extensions==3.2.3
djangorestframework==3.15.2
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==28.4.1
gunicorn==23.0.0
iniconfig==2.0.0
kombu==5.4.1
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pytest==8.3.3
pytest-django==4.9.0
pytest-env==1.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
